mage: add targz function
